@@ -1,47 +1,56 @@
 {
-  "name": "trello-clone",
+  "name": "trello-auth",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.0",
-    "@angular/cdk": "^15.2.1",
-    "@angular/common": "^15.1.0",
-    "@angular/compiler": "^15.1.0",
-    "@angular/core": "^15.1.0",
-    "@angular/forms": "^15.1.0",
-    "@angular/platform-browser": "^15.1.0",
-    "@angular/platform-browser-dynamic": "^15.1.0",
-    "@angular/router": "^15.1.0",
-    "@fortawesome/angular-fontawesome": "^0.12.1",
-    "@fortawesome/fontawesome-svg-core": "^6.2.1",
-    "@fortawesome/free-brands-svg-icons": "^6.2.1",
-    "@fortawesome/free-solid-svg-icons": "^6.2.1",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "@angular/animations": "~15.2.0",
+    "@angular/cdk": "~15.2.0",
+    "@angular/common": "~15.2.0",
+    "@angular/compiler": "~15.2.0",
+    "@angular/core": "~15.2.0",
+    "@angular/forms": "~15.2.0",
+    "@angular/platform-browser": "~15.2.0",
+    "@angular/platform-browser-dynamic": "~15.2.0",
+    "@angular/router": "~15.2.0",
+    "@fortawesome/angular-fontawesome": "~0.12.1",
+    "@fortawesome/fontawesome-svg-core": "~6.3.0",
+    "@fortawesome/free-brands-svg-icons": "~6.3.0",
+    "@fortawesome/free-solid-svg-icons": "~6.3.0",
+    "rxjs": "~7.5.0",
+    "tslib": "~2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.6",
-    "@angular/cli": "~15.1.6",
-    "@angular/compiler-cli": "^15.1.0",
-    "@tailwindcss/forms": "^0.5.3",
-    "@types/jasmine": "~4.3.0",
-    "autoprefixer": "^10.4.13",
-    "jasmine-core": "~4.5.0",
-    "karma": "~6.4.0",
+    "@angular-devkit/build-angular": "~15.2.0",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/cli": "~15.2.0",
+    "@angular/compiler-cli": "~15.2.0",
+    "@tailwindcss/forms": "~0.5.2",
+    "@types/jasmine": "~4.0.0",
+    "@typescript-eslint/eslint-plugin": "5.48.2",
+    "@typescript-eslint/parser": "5.48.2",
+    "autoprefixer": "~10.4.7",
+    "eslint": "^8.33.0",
+    "jasmine-core": "~4.1.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.7",
-    "typescript": "~4.9.4"
+    "karma-jasmine": "~5.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "postcss": "~8.4.14",
+    "tailwindcss": "~3.1.6",
+    "typescript": "~4.9.5"
   }
 }
