@@ -1,52 +1,20 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import {OverlayModule} from '@angular/cdk/overlay';
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { LoginComponent } from './pages/login/login.component';
-import { BtnComponent } from './components/btn/btn.component';
-import { BoardsComponent } from './pages/boards/boards.component';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { CdkAccordionModule } from '@angular/cdk/accordion';
-import { DragDropModule } from '@angular/cdk/drag-drop';
-import { BoardComponent } from './pages/board/board.component';
-import { DialogModule } from '@angular/cdk/dialog';
-import { TodoDialogComponent } from './components/todo-dialog/todo-dialog.component';
-import {ScrollingModule} from '@angular/cdk/scrolling';
-import { ScrollComponent} from  './pages/scroll/scroll.component';
-import { HttpClientModule} from '@angular/common/http';
-import {CdkTableModule} from '@angular/cdk/table';
-import { TableComponent} from  './pages/table/table.component';
+import { HttpClientModule } from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
 
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    LoginComponent,
-    BtnComponent,
-    BoardsComponent,
-    NavbarComponent,
-    BoardComponent,
-    TodoDialogComponent,
-    ScrollComponent,
-    TableComponent,
-
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    OverlayModule,
-    FontAwesomeModule,
-    CdkAccordionModule,
-    DragDropModule,
-    DialogModule,
-    ScrollingModule,
     HttpClientModule,
-    CdkTableModule,
     ReactiveFormsModule
-
   ],
   providers: [],
   bootstrap: [AppComponent]
