@@ -1,31 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from './pages/login/login.component';
-import { BoardsComponent } from './pages/boards/boards.component';
-import { BoardComponent } from './pages/board/board.component';
-import { ScrollComponent} from  './pages/scroll/scroll.component';
-import { TableComponent} from  './pages/table/table.component';
 
 const routes: Routes = [
   {
-    path: 'login',
-    component: LoginComponent
+    path: '',
+    loadChildren: () => import('./modules/auth/auth.module').then((m) => m.AuthModule),
   },
   {
-    path: 'boards',
-    component: BoardsComponent
-  },
-  {
-    path: 'board',
-    component: BoardComponent
-  },
-  {
-    path: 'scroll',
-    component: ScrollComponent
-  },
-  {
-    path: 'table',
-    component: TableComponent
+    path: 'app',
+    loadChildren: () => import('./modules/layout/layout.module').then((m) => m.LayoutModule),
   },
 ];
 
