@@ -1,27 +1,29 @@
 /** @type {import('tailwindcss').Config} */
-const colors = require('tailwindcss/colors')
+
+const colors = require('tailwindcss/colors');
+
 module.exports = {
-   content: [
+  content: [
     "./src/**/*.{html,ts}",
   ],
   theme: {
     extend: {
-      colors:{
+      colors: {
         success: colors.green,
         primary: colors.blue,
+        danger: colors.red,
       },
-      container:{
-        screens:{
-          sm:'640px',
-          md:'768px',
-          lg:'1024px',
-          xl:'1024px',
-          '2xl':'1536px',}
+      container: {
+        screens: {
+          sm: '640px',
+          md: '768px',
+          lg: '1024px',
+          xl: '1024px',
+          '2xl': '1536px',
+        },
       }
-
     },
   },
-
   plugins: [
     require('@tailwindcss/forms'),
   ],
